Removed "networksis" as a dependency. It has a bug I can not fix now and having it loaded by default is of marginal benefit.
Updated the citation information to be consistent with the new rules.

Added Pavel as a co-author (Hurrah, at last!!!)
Made Dave the maintainer (Hurrah, at last!!!)

It still needs to be converted to the "statnet commons" motif, but this will take until I have time to do it right.

It has been uploaded to CRAN in this form.

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@10782 9cb281f4-553f-406f-8d7f-11f072bbfe46